+++ app/templates/gulpfile.js
@@ -9,7 +9,9 @@ var $ = require('gulp-load-plugins')();
 var del = require('del');
 // 'fs' is used to read files from the system (used for AWS uploading)
 var fs = require('fs');
-// BrowserSync isn't a gulp package, and needs to be loaded manually
+<% if (amazonCloudfrontS3) { %>// Parallelize the uploads when uploading to Amazon S3
+var parallelize = require("concurrent-transform");
+<% } %>// BrowserSync isn't a gulp package, and needs to be loaded manually
 var browserSync = require('browser-sync');
 // merge is used to merge the output from two different streams into the same stream
 var merge = require('merge-stream');
@@ -105,6 +107,62 @@ gulp.task('html', ['styles'], function () {
         .pipe(gulp.dest('site'))
         .pipe($.size({title: 'Optimizations'}));
 });
+<% if (amazonCloudfrontS3) { %>
+// Task to deploy your site to Amazon S3 and Cloudfront
+gulp.task('deploy', function () {
+    // Generate the needed credentials (bucket, secret key etc) from a "hidden" JSON file
+    var credentials = JSON.parse(fs.readFileSync('aws-credentials.json', 'utf8'));
+    var publisher = $.awspublish.create(credentials);
+    // Give your files the proper headers
+    var headers = {
+        'Cache-Control': 'max-age=315360000, no-transform, public',
+    };
+
+    gulp.src('site/**/*')
+        // Gzip your files for even more zoom
+        .pipe($.awspublish.gzip())
+        // Parallelize the number of concurrent uploads, in this case 30
+        .pipe(parallelize(publisher.publish(headers), 30))
+        // Have your files in the system cache so you don't have to recheck all the files every time
+        .pipe(publisher.cache())
+        // Synchronize the contents of the bucket and local (this deletes everything that isn't in local!)
+        .pipe(publisher.sync())
+        // And print the ouput, glorious
+        .pipe($.awspublish.reporter())
+        // And update the default root object
+        .pipe($.cloudfront(credentials));
+});<% } %><% if (rsync) { %>
+// Gzip all your files before being uploaded with Rsync
+// NOTE: THIS CURRENTLY OVERWRITES THE FILES EVERY TIME SO IT'LL SYNC ALL THE FILES EVERY TIME
+// THIS IS NOT INTENDED
+gulp.task('gzip', function() {
+    // Only gzips text files and such
+    return gulp.src(['site/**/*.html', 'site/**/*.css', 'site/**/*.js', 'site/**/*.xml', 'site/**/*.txt'])
+        // Doesn't append the .gz ending to your files
+        .pipe($.gzip({append: false}))
+        .pipe(gulp.dest('site'))
+        .pipe($.size({title: 'Gzip'}));
+});
+
+// Task to upload your site via Rsync to your server
+gulp.task('deploy', ['gzip'], function () {
+    // Load in the variables needed for our Rsync synchronization
+    var secret = require('./rsync-credentials.json');
+
+    return gulp.src('site/**')
+        .pipe($.rsync({
+            // This uploads the contenst of 'root', instead of the folder
+            root: 'site',
+            // Find your username, hostname and destination from your rsync-credentials.json
+            hostname: secret.hostname,
+            username: secret.username,
+            destination: secret.destination,
+            // Incremental uploading, adds a small delay but minimizes the amount of files transferred
+            incremental: true,
+            // Shows the progress on your files while uploading
+            progress: true
+    }));
+});<% } %>
 
 // Run JS Lint against your JS
 gulp.task('jslint', function () {